crm完成

+++ workspace/crm/src/org/wang/crm/dao/impl/DepartmentDaoImpl.java
@@ -11,22 +11,28 @@ import org.wang.crm.dao.DepartmentDao;
 import org.wang.crm.dao.base.impl.CommonDaoImpl;
 import org.wang.crm.domain.Department;
 
-public class DepartmentDaoImpl extends CommonDaoImpl<Department> implements DepartmentDao {
+public class DepartmentDaoImpl extends CommonDaoImpl<Department> implements
+		DepartmentDao {
 
 	@Override
 	public void deleteByOrder() {
-		 this.getHibernateTemplate().execute(new HibernateCallback<Department>() {
+		this.getHibernateTemplate().execute(
+				new HibernateCallback<Department>() {
 
-			@Override
-			public Department doInHibernate(Session session) throws HibernateException,
-					SQLException {
-				Query query= session.createQuery("select max(did) from Department");
-				Serializable did= (Serializable) query.uniqueResult();
-				Department department=(Department) session.get(Department.class, did);
-				session.delete(department);
-				return null;
-			}
-		});
+					@Override
+					public Department doInHibernate(Session session)
+							throws HibernateException, SQLException {
+						Query query = session
+								.createQuery("select max(did) from Department");
+						Serializable did = (Serializable) query.uniqueResult();
+						if (did != null) {
+							Department department = (Department) session.get(
+									Department.class, did);
+							session.delete(department);
+						}
+						return null;
+					}
+				});
 	}
 
 }
